+++ App.js
@@ -24,6 +24,7 @@ import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { useNavigation } from "@react-navigation/native";
 import { LinearGradient } from "expo-linear-gradient";
+import LoginScreen from "./src/components/Login";
 
 const uri = "https://ak.picdn.net/shutterstock/videos/1060308725/thumb/1.jpg";
 const profilePicture = "https://randomuser.me/api/portraits/men/100.jpg";
@@ -32,160 +33,6 @@ function HomeScreen() {
   return (
     <View style={{ flex: 1, alignItems: "center", justifyContent: "center" }}>
       <Text>Home Screen</Text>
-    </View>
-  );
-}
-
-function LoginScreen() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const navigation = useNavigation();
-
-  const app = initializeApp(firebaseConfig);
-  const auth = getAuth(app);
-
-  const handleCreateAccount = () => {
-    createUserWithEmailAndPassword(auth, email, password)
-      .then((userCredential) => {
-        console.log("Account created!");
-        const user = userCredential.user;
-        console.log(user);
-      })
-      .catch((error) => {
-        console.log(error);
-        Alert.alert(error.message);
-      });
-  };
-
-  const handleSignIn = () => {
-    signInWithEmailAndPassword(auth, email, password)
-      .then((userCredential) => {
-        console.log("Signed in!");
-        const user = userCredential.user;
-        console.log(user);
-        navigation.navigate("Home");
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  };
-
-  return (
-    <View style={styles.container}>
-      <LinearGradient
-        colors={["#70E1F5", "#FFD194"]}
-        style={styles.background}
-        start={[0, 1]}
-        end={[0.75, 1]}
-      >
-        <ScrollView
-          contentContainerStyle={{
-            flex: 1,
-            width: "100%",
-            height: "100%",
-            alignItems: "center",
-            justifyContent: "center",
-          }}
-        >
-          <BlurView intensity={100}>
-            <View style={styles.login}>
-              <Image
-                source={{ uri: profilePicture }}
-                style={styles.profilePicture}
-              />
-              <View>
-                <Text
-                  style={{ fontSize: 17, fontWeight: "400", color: "white" }}
-                >
-                  E-mail
-                </Text>
-                <TextInput
-                  onChangeText={(text) => setEmail(text)}
-                  style={styles.input}
-                  placeholder="betomoedano@outlook.com"
-                />
-              </View>
-              <View>
-                <Text
-                  style={{ fontSize: 17, fontWeight: "400", color: "white" }}
-                >
-                  Password
-                </Text>
-                <TextInput
-                  onChangeText={(text) => setPassword(text)}
-                  style={styles.input}
-                  placeholder="password"
-                  secureTextEntry={true}
-                />
-              </View>
-              <TouchableOpacity
-                onPress={handleSignIn}
-                style={[styles.button, { backgroundColor: "#00CFEB90" }]}
-              >
-                <Text
-                  style={{ fontSize: 14, fontWeight: "400", color: "white" }}
-                >
-                  Login
-                </Text>
-              </TouchableOpacity>
-              <TouchableOpacity
-                onPress={handleCreateAccount}
-                style={[styles.button, { backgroundColor: "#6792F090" }]}
-              >
-                <Text
-                  style={{ fontSize: 14, fontWeight: "400", color: "white" }}
-                >
-                  Create Account
-                </Text>
-              </TouchableOpacity>
-              <View flexDirection="row">
-                <TouchableOpacity
-                  onPress={handleSignIn}
-                  style={[
-                    styles.button,
-                    styles.social,
-                    { backgroundColor: "#00CFEB90" },
-                  ]}
-                >
-                  <Text
-                    style={{ fontSize: 14, fontWeight: "400", color: "white" }}
-                  >
-                    G+
-                  </Text>
-                </TouchableOpacity>
-                <TouchableOpacity
-                  onPress={handleSignIn}
-                  style={[
-                    styles.button,
-                    styles.social,
-                    { backgroundColor: "#00CFEB90" },
-                  ]}
-                >
-                  <Text
-                    style={{ fontSize: 14, fontWeight: "400", color: "white" }}
-                  >
-                    Apple
-                  </Text>
-                </TouchableOpacity>
-                <TouchableOpacity
-                  onPress={handleSignIn}
-                  style={[
-                    styles.button,
-                    styles.social,
-                    { backgroundColor: "#00CFEB90" },
-                  ]}
-                >
-                  <Text
-                    style={{ fontSize: 14, fontWeight: "400", color: "white" }}
-                  >
-                    Fb
-                  </Text>
-                </TouchableOpacity>
-              </View>
-            </View>
-          </BlurView>
-        </ScrollView>
-      </LinearGradient>
     </View>
   );
 }
